Issues-72: Create separate module for thing diffUtils.
