+++ thingdiffutils/src/main/AndroidManifest.xml
@@ -1,0 +1,8 @@
+<manifest package="com.kirchhoff.thingdiffutils"
+	xmlns:android="http://schemas.android.com/apk/res/android">
+
+	<application>
+		<activity android:name=".ThingDiffUtilActivity" />
+	</application>
+
+</manifest>
