+++ thingdiffutils/src/main/res/layout/a_thing_diff_util.xml
