+++ thingdiffutils/src/main/res/layout/r_thing.xml
